requirement.txt
tensorflow
numpy
tqdm
pillow
scikit-learn
opencv-python 
streamlit